touchscreen: ft5x46: fix varargs warning reported with clang
drivers/input/touchscreen/ft5x46/ft5x46_ts.c:222:17: warning: passing an object that undergoes default argument promotion to 'va_start' has undefined behavior [-Wvarargs]
        va_start(varg, len);
                       ^
drivers/input/touchscreen/ft5x46/ft5x46_ts.c:212:60: note: parameter of type 'u8' (aka 'unsigned char') is declared here
static int ft5x46_recv_byte(struct ft5x46_data *ft5x46, u8 len, ...)
                                                           ^
drivers/input/touchscreen/ft5x46/ft5x46_ts.c:247:17: warning: passing an object that undergoes default argument promotion to 'va_start' has undefined behavior [-Wvarargs]
        va_start(varg, len);
                       ^
drivers/input/touchscreen/ft5x46/ft5x46_ts.c:242:60: note: parameter of type 'u8' (aka 'unsigned char') is declared here
static int ft5x46_send_byte(struct ft5x46_data *ft5x46, u8 len, ...)

+++ drivers/input/touchscreen/ft5x46/ft5x46_ts.c
@@ -209,7 +209,7 @@ static void ft5x46_cover_mode(struct ft5x46_data *ft5x46, bool enable);
 
 struct ft5x46_data *ft_data;
 
-static int ft5x46_recv_byte(struct ft5x46_data *ft5x46, u8 len, ...)
+static int ft5x46_recv_byte(struct ft5x46_data *ft5x46, int len, ...)
 {
 	int error;
 	va_list varg;
@@ -239,7 +239,7 @@ static int ft5x46_recv_block(struct ft5x46_data *ft5x46,
 	return ft5x46->bops->recv(ft5x46->dev, buf, len);
 }
 
-static int ft5x46_send_byte(struct ft5x46_data *ft5x46, u8 len, ...)
+static int ft5x46_send_byte(struct ft5x46_data *ft5x46, int len, ...)
 {
 	va_list varg;
 	u8 i, buf[len];
